requirements.txt
streamlit==1.38.0
google-generativeai==0.8.3
pandas==2.2.3
plotly==5.24.1
seaborn==0.13.2
matplotlib==3.9.2 